Move backend_connections tests
